+++ src/utils/compare.js
@@ -1,0 +1,11 @@
+export const compare = (property) => {
+  return (a, b) => {
+    if (a[property] < b[property]) {
+      return 1;
+    }
+    if (a[property] > b[property]) {
+      return -1;
+    }
+    return 0;
+  };
+};
